'cult console', prep for gemification.
Cult::NodeFile and Cult::RoleFile are now Cult::Node and Cult::Role

I've added 'cult console', that opens an IRB (or Pry or Ripl with -p or -r, respectively) with your
cult project loaded.  Cult.project will point to the right place, and some shorthand globals are set,
`$p` for project, `$r` for roles, etc.

Some file renaming for consistency has happened, and put into a structure that will make it easier to
get this gemified in an upcoming commit.